requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
bootstrap4==0.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
contextlib2==21.6.0
dj-database-url==1.2.0
dj-static==0.0.6
Django==4.1.5
django-bootstrap3==22.2
django-bootstrap4==22.3
django-stdimage==6.0.1
docopt==0.6.2
gunicorn==20.1.0
idna==3.4
mysql==0.0.3
mysqlclient==2.1.1
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
PyMySQL==1.0.2
python-dateutil==2.8.2
pythonanywhere==0.10.3
requests==2.28.2
schema==0.7.5
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
static3==0.7.0
tabulate==0.9.0
typer==0.7.0
urllib3==1.26.14
whitenoise==6.3.0
